openstack: add Ingress to infrastructure object
