+++ control-plane-operator/hostedclusterconfigoperator/controllers/resources/resources.go
@@ -652,13 +652,17 @@ func (r *reconciler) reconcileConfig(ctx context.Context, hcp *hyperv1.HostedCon
 	var currentInfra *configv1.Infrastructure
 	if _, err := r.CreateOrUpdate(ctx, r.client, infra, func() error {
 		currentInfra = infra.DeepCopy()
-		globalconfig.ReconcileInfrastructure(infra, hcp)
+		if err := globalconfig.ReconcileInfrastructure(infra, hcp); err != nil {
+			return err
+		}
 		return nil
 	}); err != nil {
 		errs = append(errs, fmt.Errorf("failed to reconcile infrastructure config spec: %w", err))
 	} else {
 		// It is reconciled a second time to update its status
-		globalconfig.ReconcileInfrastructure(infra, hcp)
+		if err := globalconfig.ReconcileInfrastructure(infra, hcp); err != nil {
+			errs = append(errs, fmt.Errorf("failed to reconcile infrastructure config status: %w", err))
+		}
 		if !equality.Semantic.DeepEqual(infra.Status, currentInfra.Status) {
 			if err := r.client.Status().Update(ctx, infra); err != nil {
 				errs = append(errs, fmt.Errorf("failed to update infrastructure status: %w", err))
