+++ hypershift-operator/controllers/nodepool/openstack/openstack.go
@@ -2,12 +2,10 @@ package openstack
 
 import (
 	"fmt"
-	"net"
 
 	hyperv1 "github.com/openshift/hypershift/api/hypershift/v1beta1"
 	"github.com/openshift/hypershift/api/util/ipnet"
 
-	"github.com/apparentlymart/go-cidr/cidr"
 	openstackutil "github.com/openshift/hypershift/support/openstackutil"
 	utilpointer "k8s.io/utils/pointer"
 	capiopenstack "sigs.k8s.io/cluster-api-provider-openstack/api/v1beta1"
@@ -39,7 +37,7 @@ func MachineTemplateSpec(hcluster *hyperv1.HostedCluster, nodePool *hyperv1.Node
 	} else {
 		machineNetworks = hcluster.Spec.Networking.MachineNetwork
 	}
-	ingressIP, err := getIngressIP(machineNetworks[0])
+	ingressIP, err := openstackutil.GetIngressIP(machineNetworks[0])
 	if err != nil {
 		return nil, err
 	}
@@ -52,20 +50,4 @@ func MachineTemplateSpec(hcluster *hyperv1.HostedCluster, nodePool *hyperv1.Node
 	openStackMachineTemplate.Template.Spec.Ports = []capiopenstack.PortOpts{port}
 
 	return openStackMachineTemplate, nil
-}
-
-// getIngressIP returns the IP address to be used for the Ingress VIP.
-// We take the seventh IP from the CIDR range like we do in the installer.
-// https://github.com/openshift/installer/blob/8e548c31b0431419350edd1fabd4dcb06263440f/pkg/types/openstack/defaults/platform.go#L48
-func getIngressIP(machineNetwork hyperv1.MachineNetworkEntry) (string, error) {
-	// go-cidr expects a net.IPNet, so we need to convert the hypershift type of CIDR to net.IPNet
-	machineNetworkIPNet := &net.IPNet{
-		IP:   machineNetwork.CIDR.IP,
-		Mask: machineNetwork.CIDR.Mask,
-	}
-	ip, err := cidr.Host(machineNetworkIPNet, 7)
-	if err != nil {
-		return "", err
-	}
-	return ip.String(), nil
 }
