--- support/openstackutil/constants.go
@@ -1,5 +1,0 @@
-package openstackutil
-
-const (
-	DefaultCIDRBlock = "10.0.0.0/16"
-)
